WIP reworking assessment hooks
